add moon and sensor to stats and commands to set/del, renamed and extracted getStatsContent
