+++ commands/stats.py
@@ -2,7 +2,7 @@ import interactions
 import logging
 from utils.authorisation import Authorization
 from utils.db import DB
-from utils.chartMaker import ChartMaker
+from utils.statsCreator import StatsCreator
 
 class Stats(interactions.Extension):
     def __init__(self, client, args):
@@ -12,7 +12,8 @@ class Stats(interactions.Extension):
         self._client: interactions.Client = client
         self._auth:Authorization = args[0]
         self._db:DB = args[1]
-        self._chartMaker:ChartMaker = args[2]
+        self._statsCreator:StatsCreator = args[2]
+        
     
     @interactions.extension_command(
         name="stats",
@@ -34,86 +35,14 @@ class Stats(interactions.Extension):
             await ctx.send(embeds=self._auth.NOT_AUTHORIZED_EMBED, ephemeral=True)
             return
         
-        try:
-            statsEmbed,statsComponent = self._getStatsContent(username)
-        except Exception as e:
-            self._logger.warning("Stats content may have an Error")
-            self._logger.warning(e)
-            await ctx.send(f"{username} nicht gefunden")
-            return
+        
+        statsEmbed,statsComponent = self._statsCreator.getStatsContent(username)
+        
+
+        #await ctx.send(f"{username} nicht gefunden")
         
         await ctx.send(embeds=statsEmbed, components=statsComponent)
 
-    def _getStatsContent(self, playerName:str):
-        playerData = self._db.getPlayerData(playerName)
-        if not playerData:
-            return False
-        
-        playerStats = self._db.getPlayerStats(playerData[1])
-        allianceData = self._db.getAllianceById(playerData[5])
-
-        currentPlayerStats = playerStats[0]
-
-        currentPlayerStats = self._getCurrentPlayerStats(playerStats[0])
-        diffData = self._getStatsDifference(playerStats)
-
-
-        #Embed Fields
-        statsFields = [
-            interactions.EmbedField(
-                inline=True,
-                name="Type",
-                value="Gesamt\nGebäude\nForschung\nFlotte\nVerteidigung\n"
-            ),
-            interactions.EmbedField(
-                inline=True,
-                name="Rang",
-                value= self._getEmbedValueString(currentPlayerStats[0], diffData[0])
-            ),
-            interactions.EmbedField(
-                inline=True,
-                name="Punkte",
-                value= self._getEmbedValueString(currentPlayerStats[1], diffData[1])
-            )
-        ]
-
-        #Add Planet Data
-        statsFields += self._getPlanetFields(playerData[1])
-
-        #Add Research Data
-        statsFields += self._getResearchFields(playerData[1])
-
-        #Create Embed
-        statsEmbed = interactions.Embed(
-            title=f"{playerData[2]}",
-            description= f"{playerData[1]}\n{allianceData[2]}", #PlayerId and Alliance Name
-            fields = statsFields,
-            timestamp=playerStats[0][19],
-            thumbnail=interactions.EmbedImageStruct(
-                url=self._chartMaker.getChartUrl(playerStats,playerData[2]),
-                height=720,
-                width=420
-            )
-        )
-
-        statsComponents = [
-            interactions.Button(
-                style=interactions.ButtonStyle.PRIMARY,
-                label='Planet Hinzufügen',
-                custom_id='btn_planet'
-            ),
-            interactions.Button(
-                style=interactions.ButtonStyle.PRIMARY,
-                label='Forschung ändern',
-                custom_id='btn_research'
-            ),
-            interactions.Button(
-                style=interactions.ButtonStyle.PRIMARY,
-                label='Allianz',
-                custom_id='btn_alliance'
-            ),
-        ]
-        return (statsEmbed, statsComponents)
 
     #Alliance Button
     @interactions.extension_component("btn_alliance")
@@ -301,32 +230,37 @@ class Stats(interactions.Extension):
     #Alliance Player Select 1
     @interactions.extension_component("allianceplayerselect1")
     async def alliancePlayerSelect1(self, ctx:interactions.ComponentContext, value): 
-        statsEmbed,statsButtons = self._getStatsContent(value[0])
-        await ctx.send(embeds=statsEmbed, components=statsButtons)
+        statsEmbed,statsButtons = self._statsCreator.getStatsContent(value[0])
+        await ctx.message.edit(embeds=statsEmbed, components=statsButtons)
+        await ctx.send()
 
     #Alliance Player Select 2
     @interactions.extension_component("allianceplayerselect2")
     async def alliancePlayerSelect2(self, ctx:interactions.ComponentContext, value):
-        statsEmbed,statsButtons = self._getStatsContent(value[0])
-        await ctx.send(embeds=statsEmbed, components=statsButtons)
+        statsEmbed,statsButtons = self._statsCreator.getStatsContent(value[0])
+        await ctx.message.edit(embeds=statsEmbed, components=statsButtons)
+        await ctx.send()
 
     #Alliance Player Select 3
     @interactions.extension_component("allianceplayerselect3")
     async def alliancePlayerSelect3(self, ctx:interactions.ComponentContext, value): 
-        statsEmbed,statsButtons = self._getStatsContent(value[0])
-        await ctx.send(embeds=statsEmbed, components=statsButtons)
+        statsEmbed,statsButtons = self._statsCreator.getStatsContent(value[0])
+        await ctx.message.edit(embeds=statsEmbed, components=statsButtons)
+        await ctx.send()
 
     #Alliance Player Select 4
     @interactions.extension_component("allianceplayerselect4")
     async def alliancePlayerSelect4(self, ctx:interactions.ComponentContext, value): 
-        statsEmbed,statsButtons = self._getStatsContent(value[0])
-        await ctx.send(embeds=statsEmbed, components=statsButtons)
+        statsEmbed,statsButtons = self._statsCreator.getStatsContent(value[0])
+        await ctx.message.edit(embeds=statsEmbed, components=statsButtons)
+        await ctx.send()
 
     #Alliance Player Select 5
     @interactions.extension_component("allianceplayerselect5")
     async def alliancePlayerSelect5(self, ctx:interactions.ComponentContext, value): 
-        statsEmbed,statsButtons = self._getStatsContent(value[0])
-        await ctx.send(embeds=statsEmbed, components=statsButtons)
+        statsEmbed,statsButtons = self._statsCreator.getStatsContent(value[0])
+        await ctx.message.edit(embeds=statsEmbed, components=statsButtons)
+        await ctx.send()
 
     def _getAllianceContent(self,allianceName:str):
         allianceData = self._db.getAlliance(allianceName)
@@ -415,134 +349,10 @@ class Stats(interactions.Extension):
         
         for player in allianceData[:5]:
             top5Fields[0].value += player[24] + "\n"
-            top5Fields[1].value += self._formatNumber(player[3]) + "\n"
-            top5Fields[2].value += self._formatNumber(player[4]) + "\n"
+            top5Fields[1].value += self._statsCreator.formatNumber(player[3]) + "\n"
+            top5Fields[2].value += self._statsCreator.formatNumber(player[4]) + "\n"
         
         return top5Fields
-
-    def _getEmbedValueString(self, data, diff):
-        #get largest number
-        maxLenghtData = len(max(data, key=len))
-
-        result = "`"
-
-        for elementData,elementDiff in zip(data,diff):
-            result += elementData.ljust(maxLenghtData) + " " + elementDiff + "\n"
-
-        return result + "`"
-
-    def _getPlanetFields(self, playerId:str):
-        planetData = self._db.getPlayerPlanets(playerId)
-
-        planetEmbeds = [
-            interactions.EmbedField(
-                name="Position",
-                value="",
-                inline=True
-            ),
-            interactions.EmbedField(
-                name="Mond",
-                value="",
-                inline=True
-            ),
-            interactions.EmbedField(
-                name="Sensor Phalanx",
-                value="",
-                inline=True
-            ),
-        ]
-
-        planetData.sort(key=lambda element: (element[2], element[3], element[4]))
-        for planet in planetData:
-            planetEmbeds[0].value += f"{planet[2]}\:{planet[3]}\:{planet[4]}\n"
-            planetEmbeds[1].value += f"-\n" #WIP
-            planetEmbeds[2].value += f"-\n" #WIP
-        
-        if not planetData:
-            planetEmbeds[0].value = "-"
-            planetEmbeds[1].value = "-"
-            planetEmbeds[2].value = "-"
-
-        return planetEmbeds
-
-    def _getResearchFields(self, playerId:str):
-
-        researchData = self._db.getResearch(playerId)
-
-        if not researchData:
-            researchData = [0,0,'-','-','-','Kein Eintrag']
-
-        researchFields= [
-            interactions.EmbedField(
-                inline=True,
-                name="Forschung",
-                value="Waffentechnik\nSchildtechnik\nRaumschiffpanzerung"
-            ),
-            interactions.EmbedField(
-                inline=True,
-                name="Level",
-                value=f"{researchData[2]}\n{researchData[3]}\n{researchData[4]}"
-            ),
-            interactions.EmbedField(
-                inline=True,
-                name="Letze Forschungs änderung: ",
-                value=f"{researchData[5]}"
-            )
-        ]
-
-        return researchFields
-
-    def _getCurrentPlayerStats(self, playerStats):
-        currentRank = [
-            self._formatNumber(playerStats[1]),
-            self._formatNumber(playerStats[5]),
-            self._formatNumber(playerStats[3]),
-            self._formatNumber(playerStats[9]),
-            self._formatNumber(playerStats[7]),
-        ]
-
-        currentScore = [
-            self._formatNumber(playerStats[2]),
-            self._formatNumber(playerStats[6]),
-            self._formatNumber(playerStats[4]),
-            self._formatNumber(playerStats[10]),
-            self._formatNumber(playerStats[8]),
-        ]
-        return [currentRank, currentScore]
-
-    def _getStatsDifference(self, playerStats):
-        if len(playerStats) == 1:
-            return [ 
-                ["n/a","n/a","n/a","n/a","n/a"],
-                ["n/a","n/a","n/a","n/a","n/a",]
-            ]
-
-        current = playerStats[0]
-        last= playerStats[1]
-            
-        diffRank = [
-            self._formatDiff(int(current[1])-int(last[1])),
-            self._formatDiff(int(current[5])-int(last[5])),
-            self._formatDiff(int(current[3])-int(last[3])),
-            self._formatDiff(int(current[9])-int(last[9])),
-            self._formatDiff(int(current[7])-int(last[7]))
-        ]
-
-        diffStats = [
-            self._formatDiff(int(current[2])-int(last[2])),
-            self._formatDiff(int(current[6])-int(last[6])),
-            self._formatDiff(int(current[4])-int(last[4])),
-            self._formatDiff(int(current[10])-int(last[10])),
-            self._formatDiff(int(current[8])-int(last[8]))
-        ]
-
-        return [diffRank, diffStats]
-
-    def _formatDiff(self, number):
-        return f"({number:+,})".replace(",",".")
-
-    def _formatNumber(self, number):
-        return f"{number:,}".replace(",",".")
 
 def setup(client, args):
     Stats(client, args)
